pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cricket-image-chatbot"
version = "0.1.0"
description = "A Streamlit-based chatbot application for cricket images"
requires-python = ">=3.9,<4.0"
dependencies = [
    "streamlit",
    "pandas",
    "numpy",
    "psycopg2-binary",
    "pgvector",
    "python-dotenv",
    "requests",
    "nltk",
    "scikit-learn",
    "langchain",
    "langchain-community",
    "langchain-core",
    "sentence-transformers",
    "huggingface-hub",
    "spacy",
    "textblob",
    "torch",
    "transformers"
]

[tool.setuptools]
packages = ["cricket_image_chatbot"]
